+++ src/TrackRowInfoVisDendrogram.js
@@ -1,5 +1,5 @@
 import React from "react";
 
 export default function TrackRowInfoVisDendrogram(props) {
-
+    return null;
 }
