+++ src/TrackRowInfoVisLink.js
@@ -1,5 +1,5 @@
 import React from "react";
 
 export default function TrackRowInfoVisLink(props) {
-
+    return null;
 }
